+++ src/commands/inputs.rs
@@ -449,7 +449,7 @@ pub async fn inputs(args: Args, config: Config) -> CommandResult<()> {
     }
     let results = Analysis::default()
         .add_source(args.source.clone())
-        .fallback_version(config.common.wdl.fallback_version)
+        .fallback_version(config.common.wdl.fallback_version.inner())
         .run()
         .await
         .map_err(CommandError::from)?;